Make next_func_idx an unsigned integer.
This prevents an integer overflow that bypasses this check: https://github.com/VirusTotal/yara/commit/f90cd22379b7efd92507f39c11fc711e3da6bebb#diff-b9cb8d707e398c02610df83e6a9a6d6a2ed0a1e9b76108b8ef45916b000f2155R616

+++ libyara/include/yara/types.h
@@ -999,7 +999,7 @@ struct YR_STRING_SET_ITERATOR
 struct YR_ITERATOR
 {
   // Index of the next function within the iter_next_func_table global array.
-  int next_func_idx;
+  uint8_t next_func_idx;
 
   union
   {
